+++ src/aks-preview/azext_aks_preview/_params.py
@@ -1318,6 +1318,7 @@ def load_arguments(self, _):
             validator=validate_node_public_ip_tags,
             help="space-separated tags: key[=value] [key[=value] ...].",
         )
+        c.argument('skip_gpu_driver_install', action='store_true', is_preview=True)
 
     with self.argument_context("aks nodepool update") as c:
         c.argument(